grpc_api/requirements.txt
# gRPC dependencies
grpcio==1.60.0
grpcio-tools==1.60.0

# Monitoring
prometheus-client==0.19.0

# ML Model (shared with other APIs)
sentence-transformers==2.3.1
torch==2.1.2
transformers==4.36.2
